+++ Package.swift
@@ -15,11 +15,16 @@ let package = Package(
         .plugin(name: "DependencyAuditPlugin", targets: ["DependencyAuditPlugin"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/apple/swift-argument-parser.git", from: "1.6.1")
+        .package(url: "https://github.com/apple/swift-argument-parser.git", from: "1.6.1"),
+        .package(url: "https://github.com/swiftlang/swift-syntax.git", from: "601.0.1"),
     ],
     targets: [
         .target(
             name: "SwiftDependencyAuditLib",
+            dependencies: [
+                .product(name: "SwiftSyntax", package: "swift-syntax"),
+                .product(name: "SwiftParser", package: "swift-syntax"),
+            ],
             plugins: ["VersionPlugin"]
         ),
         .executableTarget(